Load plugins in the main iocsh script
Loading the default plugin collection is something that is up to the
user choose to do. Otherwise, if one won't want to load one type of
plugin, they will need to stop using the whole device template. Load
that in the main script for each example template instead.

Also move loading the device template upwards, to make clearer that some
variables are only plugin specific and belong to their own logical
block.