Default to usual plugin-specific variable values
Plugins have specific definitions that can be provided when
instantiating them or loading their templates. These definitions are
hardly changed between deploys. As such, define them as the default
value to simplify main start-up scripts.

We document them in a dedicated section of variables, so that is clear
which ones are required or optional.